refactor: align contracts with subsystem boundaries
