+++ website/src/content/docs/docs/reference/index.mdx
@@ -10,4 +10,4 @@ This page indexes the four reference tables for Bub's public surface. Each page 
 | [Hooks](/docs/reference/hooks/) | Every spec in `BubHookSpecs` with kind, parameters, return type, invocation site. |
 | [CLI](/docs/reference/cli/) | Every `bub` command and subcommand, with options, defaults, and behavior notes. |
 | [Settings](/docs/reference/settings/) | All `BUB_*` environment variables, `pydantic-settings` classes, and `~/.bub/config.yml` keys. |
-| [Types](/docs/reference/types/) | Public types from `bub`: `Envelope`, `State`, `TurnResult`, runtime options, `Channel`, `OutboundChannelRouter`, `BubFramework`. |
+| [Core contracts](/docs/reference/types/) | Subsystem-owned contracts: envelopes, turns, model selection, streaming, hooks, channels, and `BubFramework`. |
